[IMP] mrp_unbuild_bom_cust_qty: total unbuild product qty automation // units coverage
With this improvement it's possible to automatically update unbuild
product quantity with the sum of every BoM quants, depending on a
general parameter that could be (de)activated.

And also, due to quantity calculation, units become important, starting
from this commit are shown in bom detailed and total lists.
Those quantities that belong to a category unit that is different than
product unbuild one are ignored during main quantity calculation and
shown in italics.