[IMP] mrp_unbuild_qc: related unbuild product support
With this improvement, an additional product for a QC inspection is
added and can be manually stablished and/or automatically filled
depending on linked stock move.
This could be useful for those inspections programmatically created and
not yet linked to an unbuild stock move (e.g. when unbuild is still in
a draft state)