[IMP] mdc_weight_mgmt_xml_import: Automatic deletion of processed files
With this imp, you can set up an automatic deletion of processed XML files
after a certain number of days. This is configured in the Maintenance Equipment form
view, where a new field "Days to Delete Processed Files" has been added. If
this field is set to a value greater than zero, a scheduled action will run daily
to delete files in the processed path that are older than the specified number of days.